Reusing instantiation

+++ src/model/entities/Department.java
@@ -6,13 +6,13 @@ public class Department implements Serializable {
 
 	private static final long serialVersionUID = 1L;
 	
-	private long id;
+	private Integer id;
 	private String name;
 	
 	public Department() {
 	}
 
-	public Department(long id, String name) {
+	public Department(Integer id, String name) {
 		this.id = id;
 		this.name = name;
 	}
@@ -21,7 +21,7 @@ public class Department implements Serializable {
 		return id;
 	}
 
-	public void setId(long id) {
+	public void setId(Integer id) {
 		this.id = id;
 	}
